+++ src/fullctl/django/context_processors.py
@@ -1,9 +1,14 @@
 from datetime import datetime
 
+import structlog
 from django.conf import settings
 
 from fullctl.django.auth import RemotePermissionsError
-from fullctl.service_bridge.aaactl import ServiceApplication
+from fullctl.django.models.concrete.account import Organization
+from fullctl.django.util import DEFAULT_FULLCTL_BRANDING
+from fullctl.service_bridge.aaactl import OrganizationBranding, ServiceApplication
+
+log = structlog.get_logger("django")
 
 
 def conf(request):
@@ -24,6 +29,7 @@ def conf(request):
 def account_service(request):
     context = {}
     org = getattr(request, "org", None)
+    context["org_branding"] = {}
 
     if org:
         org_slug = org.slug
@@ -31,6 +37,69 @@ def account_service(request):
         org_slug = ""
 
     local_auth = getattr(settings, "USE_LOCAL_PERMISSIONS", False)
+    branding_org = getattr(settings, "BRANDING_ORG", None)
+    http_host = request.get_host()
+
+    try:
+        # TODO: Look into appreach to return org specific branding or default org_branding
+        org_branding = OrganizationBranding().first(org=org_slug)
+        organization = Organization.objects.get(slug=org_slug)
+        custom_org = True
+
+        if not org_branding:
+            if branding_org:
+                org_branding = OrganizationBranding.objects.filter(
+                    org=branding_org
+                ).first()
+                if org_branding:
+                    organization = Organization.objects.get(slug=branding_org)
+                    css_dict = org_branding.css
+            elif http_host:
+                org_branding = OrganizationBranding.objects.filter(
+                    http_host=http_host
+                ).first()
+                if org_branding:
+                    organization = Organization.objects.get(slug=org_slug)
+                    css_dict = org_branding.css
+        else:
+            css_dict = org_branding.css
+
+        if org_branding and organization:
+            context["org_branding"] = {
+                "name": organization.name,
+                "html_footer": org_branding.html_footer,
+                "css": css_dict,
+                "dark_logo_url": org_branding.dark_logo_url,
+                "light_logo_url": org_branding.light_logo_url,
+                "custom_org": custom_org,
+                "show_logo": org_branding.show_logo,
+            }
+
+        if not org_branding and not branding_org and not http_host:
+            context["org_branding"] = DEFAULT_FULLCTL_BRANDING
+
+    except Exception as e:
+        log.error(f"Error fetching org org_branding: {e}")
+        context["org_branding"] = DEFAULT_FULLCTL_BRANDING
+
+    if not context["org_branding"].get("dark_logo_url", None):
+        service_logo_dark = f"{settings.SERVICE_TAG}/logo-darkbg.svg"
+    else:
+        service_logo_dark = context["org_branding"].get("dark_logo_url")
+
+    if not context["org_branding"].get("light_logo_url", None):
+        service_logo_light = f"{settings.SERVICE_TAG}/logo-lightbg.svg"
+    else:
+        service_logo_light = context["org_branding"].get("light_logo_url")
+
+    if not context["org_branding"].get("name", None):
+        logo_alt_text = settings.SERVICE_TAG
+        service_name = settings.SERVICE_TAG.replace("ctl", "")
+    else:
+        logo_alt_text = context["org_branding"].get("name")
+        service_name = context["org_branding"].get("name")
+
+    service_tag = settings.SERVICE_TAG
 
     # TODO abstract so other auth services can be
     # defined
@@ -45,16 +114,21 @@ def account_service(request):
             },
         },
         oauth_manages_org=not local_auth,
-        service_logo_dark=f"{settings.SERVICE_TAG}/logo-darkbg.svg",
-        service_logo_light=f"{settings.SERVICE_TAG}/logo-lightbg.svg",
-        service_tag=settings.SERVICE_TAG,
-        service_name=settings.SERVICE_TAG.replace("ctl", ""),
+        service_logo_dark=service_logo_dark,
+        service_logo_light=service_logo_light,
+        service_tag=service_tag,
+        service_name=service_name,
+        logo_alt_text=logo_alt_text,
     )
 
     if settings.OAUTH_TWENTYC_URL:
         context.update(
             service_applications=[
-                service_application.for_org(org)
+                # we call sanitize() to remove the `config` attribute
+                # as that may contain sensitive information
+                # that we don't necessarily want to expose to the template
+                # context
+                service_application.for_org(org).sanitize()
                 for service_application in ServiceApplication().objects(
                     group="fullctl", org=(org_slug or None)
                 )
@@ -73,11 +147,13 @@ def account_service(request):
 
     if local_auth:
         context["service_info"] = {
-            "name": settings.SERVICE_TAG,
+            "name": f"{settings.SERVICE_TAG} {context['org_branding']['name']}"
+            if context["org_branding"].get("name", None)
+            else settings.SERVICE_TAG,
             "slug": settings.SERVICE_TAG,
             "description": "Local permissions",
             "org_has_access": True,
-            "org_namespace": f"{settings.SERVICE_TAG}",
+            "org_namespace": settings.SERVICE_TAG,
         }
 
     return context